backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/narrative_mirror"]

[project]
name = "narrarc"
version = "0.1.0"
description = "Narrative Mirror - Transform WeChat chat history into evidence-backed narrative arcs"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.5.0",
    "httpx[socks]>=0.28.1",
    "langgraph>=0.2,<0.4",
    "openai>=2.21.0",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pyyaml>=6.0.3",
]

[tool.pytest.ini_options]
markers = [
    "integration: real API calls, requires config.yml",
    "dataset: requires locally-downloaded public dataset (see REALTALK_PATH, KAGGLE_WA_CSV, CANDOR_SESSION_JSON env vars)",
]
